llm-server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
gunicorn==23.0.0
openai>=1.35.0,<2.0.0
websockets==13.1
tiktoken==0.8.0
python-dotenv==1.0.1
numpy==1.26.2
websocket-client==1.8.0

# MongoDB
pymongo==4.9.0
motor==3.6.0

# Authentication - Using Argon2 (modern best practice)
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.19

# OAuth
authlib==1.3.2
httpx==0.28.1
itsdangerous==2.2.0

# Validation - Pydantic v2
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Date handling
python-dateutil==2.9.0